Drop hb-ot-color-colr-table.cc
Move everything into the .hh file.
