+++ src/Makefile.sources
@@ -70,7 +70,6 @@ HB_BASE_sources = \
 	hb-ot-cmap-table.hh \
 	hb-ot-color-cbdt-table.hh \
 	hb-ot-color-colr-table.hh \
-	hb-ot-color-colr-table.cc \
 	hb-ot-color-cpal-table.hh \
 	hb-ot-color-sbix-table.hh \
 	hb-ot-color-svg-table.hh \
