+++ src/harfbuzz-subset.cc
@@ -14,7 +14,6 @@
 #include "hb-number.cc"
 #include "hb-ot-cff1-table.cc"
 #include "hb-ot-cff2-table.cc"
-#include "hb-ot-color-colr-table.cc"
 #include "hb-ot-color.cc"
 #include "hb-ot-face.cc"
 #include "hb-ot-font.cc"
